+++ lib/contract.ex
@@ -4,21 +4,32 @@ defmodule Soroban.Contract do
   """
 
   alias Soroban.Contract.{
+    BumpFootprintExpiration,
     DeployAssetContract,
     DeployContract,
     InvokeContractFunction,
+    RestoreFootprint,
     UploadContractCode
   }
 
   defdelegate invoke(
-                contract_id,
+                contract_address,
                 source_secret_key,
                 function_name,
                 function_args \\ [],
-                auth_secret_key \\ nil
+                auth_secret_keys \\ []
               ),
               to: InvokeContractFunction,
               as: :invoke
+
+  defdelegate simulate_invoke(
+                contract_address,
+                source_public_key,
+                function_name,
+                function_args \\ []
+              ),
+              to: InvokeContractFunction,
+              as: :simulate_invoke
 
   defdelegate upload(
                 wasm,
@@ -41,8 +52,32 @@ defmodule Soroban.Contract do
               to: DeployAssetContract,
               as: :deploy
 
+  defdelegate bump_contract(contract_address, secret_key, ledgers_to_bump),
+    to: BumpFootprintExpiration,
+    as: :bump_contract
+
+  defdelegate bump_contract_wasm(wasm_id, secret_key, ledgers_to_bump),
+    to: BumpFootprintExpiration,
+    as: :bump_contract_wasm
+
+  defdelegate bump_contract_keys(contract_address, secret_key, ledgers_to_bump, keys),
+    to: BumpFootprintExpiration,
+    as: :bump_contract_keys
+
+  defdelegate restore_contract(contract_address, secret_key),
+    to: RestoreFootprint,
+    as: :restore_contract
+
+  defdelegate restore_contract_wasm(wasm_id, secret_key),
+    to: RestoreFootprint,
+    as: :restore_contract_wasm
+
+  defdelegate restore_contract_keys(contract_address, secret_key, keys),
+    to: RestoreFootprint,
+    as: :restore_contract_keys
+
   defdelegate retrieve_unsigned_xdr_to_invoke(
-                contract_id,
+                contract_address,
                 source_public_key,
                 function_name,
                 function_args \\ []