+++ lib/contract/invoke_contract_function.ex
@@ -4,7 +4,8 @@ defmodule Soroban.Contract.InvokeContractFunction do
   """
 
   alias Soroban.Contract.RPCCalls
-  alias Soroban.RPC.SendTransactionResponse
+  alias Soroban.RPC.{SendTransactionResponse, SimulateTransactionResponse}
+  alias Soroban.Types.Address
 
   alias Stellar.Horizon.Accounts
 
@@ -12,7 +13,6 @@ defmodule Soroban.Contract.InvokeContractFunction do
     Account,
     HostFunction,
     InvokeHostFunction,
-    SCAddress,
     SCVal,
     SCVec,
     SequenceNumber,
@@ -21,13 +21,14 @@ defmodule Soroban.Contract.InvokeContractFunction do
 
   @type account :: Account.t()
   @type function_args :: list(struct())
-  @type auth_secret_key :: String.t() | nil
+  @type auth_secret_keys :: list(String.t())
   @type invoke_host_function :: InvokeHostFunction.t()
   @type envelope_xdr :: String.t()
   @type function_name :: String.t()
   @type contract_address :: String.t()
   @type source_secret_key :: String.t()
   @type source_public_key :: String.t()
+  @type simulate_response :: {:ok, SimulateTransactionResponse.t()}
   @type send_response :: {:ok, SendTransactionResponse.t()}
   @type signature :: Signature.t()
   @type sequence_number :: SequenceNumber.t()
@@ -38,14 +39,14 @@ defmodule Soroban.Contract.InvokeContractFunction do
           source_secret_key :: source_secret_key(),
           function_name :: function_name(),
           function_args :: function_args(),
-          auth_secret_key :: auth_secret_key()
+          auth_secret_keys :: auth_secret_keys()
         ) :: send_response()
   def invoke(
         contract_address,
         source_secret_key,
         function_name,
         function_args,
-        auth_secret_key \\ nil
+        auth_secret_keys \\ []
       ) do
     with {public_key, _secret} = keypair <- Stellar.KeyPair.from_secret_seed(source_secret_key),
          {:ok, seq_num} <- Accounts.fetch_next_sequence_number(public_key),
@@ -62,7 +63,7 @@ defmodule Soroban.Contract.InvokeContractFunction do
         sequence_number,
         signature,
         invoke_host_function_op,
-        auth_secret_key
+        auth_secret_keys
       )
     end
   end
@@ -96,6 +97,33 @@ defmodule Soroban.Contract.InvokeContractFunction do
     end
   end
 
+  @spec simulate_invoke(
+          contract_address :: contract_address(),
+          source_public_key :: source_public_key(),
+          function_name :: function_name(),
+          function_args :: function_args()
+        ) :: simulate_response()
+  def simulate_invoke(
+        contract_address,
+        source_public_key,
+        function_name,
+        function_args
+      ) do
+    with {:ok, seq_num} <- Accounts.fetch_next_sequence_number(source_public_key),
+         {:ok, function_args} <- convert_to_sc_val(function_args),
+         %Account{} = source_account <- Account.new(source_public_key),
+         %SequenceNumber{} = sequence_number <- SequenceNumber.new(seq_num),
+         %InvokeHostFunction{} = invoke_host_function_op <-
+           create_host_function_op(
+             contract_address,
+             function_name,
+             function_args,
+             source_public_key
+           ) do
+      RPCCalls.simulate(invoke_host_function_op, source_account, sequence_number)
+    end
+  end
+
   @spec create_host_function_op(
           contract_address :: contract_address(),
           function_name :: function_name(),
@@ -105,8 +133,8 @@ defmodule Soroban.Contract.InvokeContractFunction do
   defp create_host_function_op(contract_address, function_name, function_args, source_public_key) do
     contract_address =
       contract_address
-      |> SCAddress.new()
-      |> (&SCVal.new(address: &1)).()
+      |> Address.new()
+      |> Address.to_sc_val()
 
     function_name = SCVal.new(symbol: function_name)
     function_args = SCVec.new([contract_address, function_name] ++ function_args)