+++ test/contract/invoke_contract_function_test.exs
@@ -42,6 +42,26 @@ defmodule Stellar.Horizon.Client.CannedAccountRequests do
       ) do
     {:ok, 200, [], "{\"sequence\":\"1390916568875069\"}"}
   end
+
+  def request(
+        :get,
+        @base_url <> "/accounts/GCODO4TKGGIBN2EWNFXTJADNYIQOPQ2XUO6IBSCHA2NM4CGTKL2TJTRE",
+        _headers,
+        _body,
+        _opts
+      ) do
+    {:ok, 200, [], "{\"sequence\":\"1390916568875069\"}"}
+  end
+
+  def request(
+        :get,
+        @base_url <> "/accounts/GCJFPGZINFE3WI6PAKCT42OMI35UMTXZZ6DT6VZOSEIK3YORRXWDEGOM",
+        _headers,
+        _body,
+        _opts
+      ) do
+    {:ok, 200, [], "{\"sequence\":\"1390916568875069\"}"}
+  end
 end
 
 defmodule Soroban.RPC.CannedInvokeContractFunctionClientImpl do
@@ -99,9 +119,7 @@ defmodule Soroban.RPC.CannedInvokeContractFunctionClientImpl do
        min_resource_fee: "79488",
        results: [
          %{
-           auth: [
-             "AAAAAAAAAAAAAAAB9naqmyRmTCwhDWcnYJzK9ZtX7TrgjMYqFwQWszcX0SAAAAADaW5jAAAAAAIAAAASAAAAAAAAAADJTnjlYcJxSKDat78jbEyDkVqo2uhpNX4BdBtqGrx+1gAAAAkAAAAAAAAAAAAAAAAAAAACAAAAAA=="
-           ],
+           auth: nil,
            xdr: "AAAAEAAAAAEAAAACAAAADwAAAAVIZWxsbwAAAAAAAA8AAAAFd29ybGQAAAA="
          }
        ],
@@ -161,6 +179,69 @@ defmodule Soroban.RPC.CannedInvokeContractFunctionClientImpl do
      }}
   end
 
+  def request(
+        "simulateTransaction",
+        _url,
+        _headers,
+        %{
+          transaction:
+            "AAAAAgAAAACcN3JqMZAW6JZpbzSAbcIg58NXo7yAyEcGms4I01L1NAAAAGQABPEIAAAAPgAAAAAAAAAAAAAAAQAAAAEAAAAAnDdyajGQFuiWaW80gG3CIOfDV6O8gMhHBprOCNNS9TQAAAAYAAAAAAAAAAMAAAASAAAAAfZ2qpskZkwsIQ1nJ2CcyvWbV+064IzGKhcEFrM3F9EgAAAADwAAAA1mdW5jdGlvbl9uYW1lAAAAAAAADwAAAANBcmcAAAAAAAAAAAAAAAAA"
+        },
+        _opts
+      ) do
+    send(self(), {:request, "RESPONSE"})
+
+    {:ok,
+     %{
+       transaction_data:
+         "AAAAAAAAAAEAAAAHmDXys1KuBimD87u2AiUG/jb5CqOkQW/qASpb6gMVRlsAAAAAAAAAAQAAAAYAAAAB9naqmyRmTCwhDWcnYJzK9ZtX7TrgjMYqFwQWszcX0SAAAAAUAAAAAQAAAAAAN4HKAAAU9AAAAWAAAAPUAAAAAAAAAMA=",
+       events: nil,
+       min_resource_fee: "79488",
+       results: [
+         %{
+           auth: [
+             "AAAAAQAAAAAAAAAAyU545WHCcUig2re/I2xMg5FaqNroaTV+AXQbahq8ftZXNpultrwTUAAAAAAAAAAAAAAAAAAAAAH5tV8LM6MblBMlB7cp0DcehSyr70sSTZsTzo1ahdNYVAAAAARzd2FwAAAAAgAAAAoAAAAAAAAAAAAAAAAAAABkAAAACgAAAAAAAAAAAAAAAAAAEZQAAAAA"
+           ],
+           xdr: "AAAAEAAAAAEAAAACAAAADwAAAAVIZWxsbwAAAAAAAA8AAAAFd29ybGQAAAA="
+         }
+       ],
+       cost: %{cpu_insns: "1048713", mem_bytes: "1201148"},
+       latest_ledger: "475528"
+     }}
+  end
+
+  def request(
+        "simulateTransaction",
+        _url,
+        _headers,
+        %{
+          transaction:
+            "AAAAAgAAAACSV5soaUm7I88ChT5pzEb7Rk75z4c/Vy6REK3h0Y3sMgAAAGQABPEIAAAAPgAAAAAAAAAAAAAAAQAAAAEAAAAAklebKGlJuyPPAoU+acxG+0ZO+c+HP1cukRCt4dGN7DIAAAAYAAAAAAAAAAMAAAASAAAAAfZ2qpskZkwsIQ1nJ2CcyvWbV+064IzGKhcEFrM3F9EgAAAADwAAAA1mdW5jdGlvbl9uYW1lAAAAAAAADwAAAANBcmcAAAAAAAAAAAAAAAAA"
+        },
+        _opts
+      ) do
+    send(self(), {:request, "RESPONSE"})
+
+    {:ok,
+     %{
+       transaction_data:
+         "AAAAAAAAAAEAAAAHmDXys1KuBimD87u2AiUG/jb5CqOkQW/qASpb6gMVRlsAAAAAAAAAAQAAAAYAAAAB9naqmyRmTCwhDWcnYJzK9ZtX7TrgjMYqFwQWszcX0SAAAAAUAAAAAQAAAAAAN4HKAAAU9AAAAWAAAAPUAAAAAAAAAMA=",
+       events: nil,
+       min_resource_fee: "79488",
+       results: [
+         %{
+           auth: [
+             "AAAAAQAAAAAAAAAAyU545WHCcUig2re/I2xMg5FaqNroaTV+AXQbahq8ftZXNpultrwTUAAAAAAAAAAAAAAAAAAAAAH5tV8LM6MblBMlB7cp0DcehSyr70sSTZsTzo1ahdNYVAAAAARzd2FwAAAAAgAAAAoAAAAAAAAAAAAAAAAAAABkAAAACgAAAAAAAAAAAAAAAAAAEZQAAAAA",
+             "AAAAAQAAAAAAAAAAwH2t7Rb1TR1p5RQ5G0eSW+7ebMqBlCOZ41wClsmaNPoW3+w0mXsgEwAAAAAAAAAAAAAAAAAAAAH5tV8LM6MblBMlB7cp0DcehSyr70sSTZsTzo1ahdNYVAAAAARzd2FwAAAAAgAAAAoAAAAAAAAAAAAAAAAAABGUAAAACgAAAAAAAAAAAAAAAAAAAGQAAAAA"
+           ],
+           xdr: "AAAAEAAAAAEAAAACAAAADwAAAAVIZWxsbwAAAAAAAA8AAAAFd29ybGQAAAA="
+         }
+       ],
+       cost: %{cpu_insns: "1048713", mem_bytes: "1201148"},
+       latest_ledger: "475528"
+     }}
+  end
+
   @impl true
   def request("sendTransaction", _url, _headers, _body, _opts) do
     send(self(), {:request, "RESPONSE"})
@@ -171,6 +252,18 @@ defmodule Soroban.RPC.CannedInvokeContractFunctionClientImpl do
        hash: "a4721e2a61e9a6b3f54030396e41c3e352101e6cd649b4453e89fb3e827744f4",
        latest_ledger: "476420",
        latest_ledger_close_time: "1683150612"
+     }}
+  end
+
+  @impl true
+  def request("getLatestLedger", _url, _headers, _body, _opts) do
+    send(self(), {:request, "RESPONSE"})
+
+    {:ok,
+     %{
+       id: "08f19981546ea2ef3ea0523a5e4156942f5b34c432b70d51ceb77e6d8e80433c",
+       protocol_version: "20",
+       sequence: 214_971
      }}
   end
 end
@@ -199,24 +292,28 @@ defmodule Soroban.Contract.InvokeContractFunctionTest do
     end)
 
     %{
-      contract_id: "CD3HNKU3ERTEYLBBBVTSOYE4ZL2ZWV7NHLQIZRRKC4CBNMZXC7ISBXHV",
+      contract_address: "CD3HNKU3ERTEYLBBBVTSOYE4ZL2ZWV7NHLQIZRRKC4CBNMZXC7ISBXHV",
       source_public: "GDEU46HFMHBHCSFA3K336I3MJSBZCWVI3LUGSNL6AF2BW2Q2XR7NNAPM",
       # GBNDWIM7DPYZJ2RLJ3IESXBIO4C2SVF6PWZXS3DLODJSBQWBMKY5U4M3
       source_secret: "SDRD4CSRGPWUIPRDS5O3CJBNJME5XVGWNI677MZDD4OD2ZL2R6K5IQ24",
       # GDDZSR7Y6TIMSBM72WYVGUH6FB6P7MF6Y6DU7MCNAPFRXI5GCWGWWFRS
       source_secret_with_auth: "SCFQVIK6JH2NNVWVZFQBA7FRKPYHIAGOMMGO32RZKTQ3QUTLU5DN67MG",
+      # GCODO4TKGGIBN2EWNFXTJADNYIQOPQ2XUO6IBSCHA2NM4CGTKL2TJTRE
+      source_secret_with_auths: "SAXEQ7CYV5MA3RUCTWWSUBPA3SVZYT2IDXO5WPYRI34POHUBOBDPVYPT",
+      # GCJFPGZINFE3WI6PAKCT42OMI35UMTXZZ6DT6VZOSEIK3YORRXWDEGOM
+      source_secret_auths_error: "SDDDMA6TT5FYTI33HBFLDCWJPJL5QFZU53B3246XYGZQ3EOQX74TFQ4I",
       source_public_with_error: "GASY52GNGVKEMXSGH7VSCZQKRWQMIQD77J53KHXEBAV2BODWH6FDDZ3F",
       source_secret_with_error: "SDXKY6TSBNS7T2UJMHLIH4BWTP4EHR52HZTRNEKH33ML3ARJI2AKIPEC",
       function_name: "function_name",
       function_args: [Symbol.new("Arg")],
-      auth_secret_key: "SCAVFA3PI3MJLTQNMXOUNBSEUOSY66YMG3T2KCQKLQBENNVLVKNPV3EK",
+      auth_secret_keys: ["SCAVFA3PI3MJLTQNMXOUNBSEUOSY66YMG3T2KCQKLQBENNVLVKNPV3EK"],
       xdr_envelope:
-        "AAAAAgAAAADJTnjlYcJxSKDat78jbEyDkVqo2uhpNX4BdBtqGrx+1gABOf8ABPEIAAAAPgAAAAAAAAAAAAAAAQAAAAEAAAAAyU545WHCcUig2re/I2xMg5FaqNroaTV+AXQbahq8ftYAAAAYAAAAAAAAAAMAAAASAAAAAfZ2qpskZkwsIQ1nJ2CcyvWbV+064IzGKhcEFrM3F9EgAAAADwAAAA1mdW5jdGlvbl9uYW1lAAAAAAAADwAAAANBcmcAAAAAAAAAAAEAAAAAAAAAAgAAAAYAAAAB9naqmyRmTCwhDWcnYJzK9ZtX7TrgjMYqFwQWszcX0SAAAAAUAAAAAQAAAAAAAAAHmDXys1KuBimD87u2AiUG/jb5CqOkQW/qASpb6gMVRlsAAAAAAAAAAAA1i+AAABQ4AAAAAAAAAPAAAAAAAAAALwAAAAA="
+        "AAAAAgAAAADJTnjlYcJxSKDat78jbEyDkVqo2uhpNX4BdBtqGrx+1gABRmoABPEIAAAAPgAAAAAAAAAAAAAAAQAAAAEAAAAAyU545WHCcUig2re/I2xMg5FaqNroaTV+AXQbahq8ftYAAAAYAAAAAAAAAAMAAAASAAAAAfZ2qpskZkwsIQ1nJ2CcyvWbV+064IzGKhcEFrM3F9EgAAAADwAAAA1mdW5jdGlvbl9uYW1lAAAAAAAADwAAAANBcmcAAAAAAAAAAAEAAAAAAAAAAgAAAAYAAAAB9naqmyRmTCwhDWcnYJzK9ZtX7TrgjMYqFwQWszcX0SAAAAAUAAAAAQAAAAAAAAAHmDXys1KuBimD87u2AiUG/jb5CqOkQW/qASpb6gMVRlsAAAAAAAAAAAA1i+AAABQ4AAAAAAAAAPAAAAAAAAAALwAAAAA="
     }
   end
 
   test "invoke host function without authorization", %{
-    contract_id: contract_id,
+    contract_address: contract_address,
     source_secret: source_secret,
     function_name: function_name,
     function_args: function_args
@@ -230,7 +327,7 @@ defmodule Soroban.Contract.InvokeContractFunctionTest do
        error_result_xdr: nil
      }} =
       InvokeContractFunction.invoke(
-        contract_id,
+        contract_address,
         source_secret,
         function_name,
         function_args
@@ -238,7 +335,7 @@ defmodule Soroban.Contract.InvokeContractFunctionTest do
   end
 
   test "invoke host function without signed authorization", %{
-    contract_id: contract_id,
+    contract_address: contract_address,
     source_secret_with_auth: source_secret_with_auth,
     function_name: function_name,
     function_args: function_args
@@ -252,15 +349,87 @@ defmodule Soroban.Contract.InvokeContractFunctionTest do
        error_result_xdr: nil
      }} =
       InvokeContractFunction.invoke(
-        contract_id,
+        contract_address,
         source_secret_with_auth,
         function_name,
         function_args
       )
   end
 
+  test "invoke host function with signed authorization", %{
+    contract_address: contract_address,
+    source_secret_with_auths: source_secret_with_auths,
+    function_name: function_name,
+    function_args: function_args,
+    auth_secret_keys: auth_secret_keys
+  } do
+    {:ok,
+     %SendTransactionResponse{
+       status: "PENDING",
+       hash: "a4721e2a61e9a6b3f54030396e41c3e352101e6cd649b4453e89fb3e827744f4",
+       latest_ledger: "476420",
+       latest_ledger_close_time: "1683150612",
+       error_result_xdr: nil
+     }} =
+      InvokeContractFunction.invoke(
+        contract_address,
+        source_secret_with_auths,
+        function_name,
+        function_args,
+        auth_secret_keys
+      )
+  end
+
+  test "simulate invoke", %{
+    contract_address: contract_address,
+    source_public: source_public,
+    function_name: function_name,
+    function_args: function_args
+  } do
+    {:ok,
+     %SimulateTransactionResponse{
+       transaction_data:
+         "AAAAAAAAAAIAAAAGAAAAAfZ2qpskZkwsIQ1nJ2CcyvWbV+064IzGKhcEFrM3F9EgAAAAFAAAAAEAAAAAAAAAB5g18rNSrgYpg/O7tgIlBv42+QqjpEFv6gEqW+oDFUZbAAAAAAAAAAAANYvgAAAUOAAAAAAAAADwAAAAAAAAAC8=",
+       events: nil,
+       min_resource_fee: "79488",
+       results: [
+         %{
+           auth: nil,
+           events: nil,
+           xdr: "AAAAEAAAAAEAAAACAAAADwAAAAVIZWxsbwAAAAAAAA8AAAAFd29ybGQAAAA="
+         }
+       ],
+       cost: %{cpu_insns: "1052105", mem_bytes: "1201148"},
+       latest_ledger: "690189",
+       error: nil
+     }} =
+      InvokeContractFunction.simulate_invoke(
+        contract_address,
+        source_public,
+        function_name,
+        function_args
+      )
+  end
+
+  test "invoke host function invalid length of auth keys", %{
+    contract_address: contract_address,
+    source_secret_auths_error: source_secret_auths_error,
+    function_name: function_name,
+    function_args: function_args,
+    auth_secret_keys: auth_secret_keys
+  } do
+    {:error, :invalid_auth_secret_keys_length} =
+      InvokeContractFunction.invoke(
+        contract_address,
+        source_secret_auths_error,
+        function_name,
+        function_args,
+        auth_secret_keys
+      )
+  end
+
   test "invoke host function with simulate error", %{
-    contract_id: contract_id,
+    contract_address: contract_address,
     source_secret_with_error: source_secret_with_error,
     function_name: function_name,
     function_args: function_args
@@ -270,7 +439,7 @@ defmodule Soroban.Contract.InvokeContractFunctionTest do
        error: "error"
      }} =
       InvokeContractFunction.invoke(
-        contract_id,
+        contract_address,
         source_secret_with_error,
         function_name,
         function_args
@@ -278,7 +447,7 @@ defmodule Soroban.Contract.InvokeContractFunctionTest do
   end
 
   test "retrieve_unsigned_xdr_to_invoke without authorization", %{
-    contract_id: contract_id,
+    contract_address: contract_address,
     source_public: source_public,
     function_name: function_name,
     function_args: function_args,
@@ -286,7 +455,7 @@ defmodule Soroban.Contract.InvokeContractFunctionTest do
   } do
     ^xdr_envelope =
       InvokeContractFunction.retrieve_unsigned_xdr_to_invoke(
-        contract_id,
+        contract_address,
         source_public,
         function_name,
         function_args
@@ -294,7 +463,7 @@ defmodule Soroban.Contract.InvokeContractFunctionTest do
   end
 
   test "retrieve_unsigned_xdr_to_invoke host function with simulate error", %{
-    contract_id: contract_id,
+    contract_address: contract_address,
     source_public_with_error: source_public_with_error,
     function_name: function_name,
     function_args: function_args
@@ -304,7 +473,7 @@ defmodule Soroban.Contract.InvokeContractFunctionTest do
        error: "error"
      }} =
       InvokeContractFunction.retrieve_unsigned_xdr_to_invoke(
-        contract_id,
+        contract_address,
         source_public_with_error,
         function_name,
         function_args