+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.13.0 (08.08.2023)
+
+- [Allow invocation with invoker authorization](https://github.com/kommitters/soroban.ex/issues/103).
+- [Add BumpFootprintExpiration operation](https://github.com/kommitters/soroban.ex/issues/104).
+- [Add RestoreFootprint operation](https://github.com/kommitters/soroban.ex/issues/105).
+- Allow to use simulate tx in contract functions invocation.
+- Update all dependencies.
+
 ## 0.12.0 (28.07.2023)
 
 - [Soroban Preview 10 support: Enable key use cases for deployment and invocation](https://github.com/kommitters/soroban.ex/issues/102).